+++ changelog.md
@@ -6,6 +6,7 @@
  - Constellation is now paid
  - Improve DNS Challenge with smarter resolution for faster and more reliable results (especially when using local nameservers)
  - Fix issues where it was impossible to login with insecure local IPs
+ - Better suppoer for container/service network_mode when importing compose
  - Default networks to 16 Ips instead of 8
  - Further improving the docker-compose import to mimic naming and hostnaming convention
  - Added hostname stickiness to compose network namespaces